rename to cli_util, refactor status, move grouping to openfoamproject
